+++ src/main/java/org/fundacionjala/sfdc/CommonActions.java
@@ -1,5 +1,6 @@
 package org.fundacionjala.sfdc;
 
+import org.fundacionjala.sfdc.driver.DriverManager;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.support.ui.ExpectedConditions;
