+++ src/main/java/org/fundacionjala/sfdc/driver/DriverType.java
@@ -1,0 +1,9 @@
+package org.fundacionjala.sfdc.driver;
+
+/**
+ * Created by abelb on 8/12/2017.
+ */
+public enum DriverType {
+    CHROME,
+    FIREFOX
+}
